Moved authentication.
