--- Application/Authentication/IAuthenticationService.cs
@@ -1,12 +1,0 @@
-﻿using Wallet.Shared.Authentication;
-
-namespace Wallet.Application.Authentication;
-
-public interface IAuthenticationService
-{
-    Task<AuthenticationResponse> LoginAsync(LoginRequest request, CancellationToken cancellationToken);
-
-    Task<AuthenticationResponse> RefreshTokenAsync(RefreshTokenRequest request, CancellationToken cancellationToken);
-
-    Task RegisterAsync(RegistrationRequest request);
-}
