--- Application/Common/Exceptions/BadRequestException.cs
@@ -1,25 +1,0 @@
-namespace Wallet.Application.Common.Exceptions;
-
-public class BadRequestException : Exception
-{
-    public BadRequestException(IEnumerable<string> errors)
-    {
-        Errors = errors.ToArray();
-    }
-
-    public BadRequestException()
-    {
-    }
-
-    public BadRequestException(string message)
-        : base(message)
-    {
-    }
-
-    public BadRequestException(string message, Exception innerException)
-        : base(message, innerException)
-    {
-    }
-
-    public IEnumerable<string> Errors { get; }
-}
