--- WebApi/Services/CurrentUserService.cs
@@ -1,16 +1,0 @@
-using System.Security.Claims;
-using Wallet.Application.Common.Interfaces;
-
-namespace Wallet.WebApi.Services;
-
-public class CurrentUserService : ICurrentUserService
-{
-    private readonly IHttpContextAccessor _httpContextAccessor;
-
-    public CurrentUserService(IHttpContextAccessor httpContextAccessor)
-    {
-        _httpContextAccessor = httpContextAccessor;
-    }
-
-    public string UserId => _httpContextAccessor.HttpContext?.User.FindFirstValue("userid");
-}
